feat: Refactor URL structure for articles, update BlogCard and BlogList components, and enhance homepage layout
