+++ .github/memory-bank/activeContext.md
@@ -1,3 +1,37 @@
+## 2025-05-15: URL構造の改善
+
+### 実装内容と成果
+
+#### URLs構造の整理
+- 直感的でないURL構造を改善
+  - 問題点: `/` と `/blog` が同じコンテンツを表示し、個別記事のURLは `/blog/[slug]` だった
+  - 解決策: 
+    - `/` をサイトの「顔」として特別なホームページに再設計
+    - `/blog` を新着記事一覧ページとして明確にポジショニング
+    - 個別記事のURLを `/articles/[slug]` に変更
+  - 実装方法:
+    - `src/app/articles/[slug]` ディレクトリを新規作成
+    - `src/app/page.tsx` を特徴的なトップページとして再設計
+    - `BlogCard` コンポーネントを更新してリンク先を更新
+      - 引数の形式を変更 (post object -> 展開されたプロパティ)
+      - デフォルトhrефを `/articles/${slug}` に設定
+    - `TagCloud` コンポーネントのプロパティ名を修正 (`maxItems` -> `limit`)
+    - テストも更新して新しいURL構造に対応
+
+#### ホームページの強化
+- トップページを特徴的な「玄関」として再設計
+  - 新着記事のプレビュー（最新5件）
+  - 人気タグへの簡単アクセス
+  - サイト紹介と各セクションへのナビゲーション
+- コンパイルエラーを修正
+  - BlogCardコンポーネントへのprops渡し方を修正
+  - TagCloudコンポーネントのプロパティ名を修正
+
+### 次のステップ
+- 時間経過後に、SEO観点での新しいURL構造の効果測定
+- ユーザービヘイビアデータの収集（各ページからの遷移率など）
+- 改善されたURLの使用状況に関するフィードバック収集
+
 ## 2025-05-15: OG画像生成のフォント読み込み改善とクリーンアップ
 
 ### 問題点と解決
