+++ .github/memory-bank/progress.md
@@ -1,5 +1,31 @@
 # Progress Tracking
 
+## 2025-05-15 URL構造の改善
+
+- ブログのURL構造を直感的で意味のある形に整理
+  - 問題点: `/` と `/blog` が同じコンテンツ（最新記事）を表示、個別記事は `/blog/[slug]` という階層的に一貫性のないパス
+  - 解決策:
+    - ホームページ（`/`）を特別な「玄関」として再設計（最新記事プレビュー、人気タグ、サイト紹介）
+    - 新着記事ページ（`/blog`）を時系列記事一覧として明確にポジショニング
+    - 個別記事のURLを `/articles/[slug]` に変更し、階層構造の一貫性を向上
+  - 実装手順:
+    - `src/app/articles/[slug]/page.tsx` を新規作成し、既存の記事ページ機能を移行
+    - ホームページを機能豊富な構造に再設計（最新記事のプレビュー、タグクラウド、サイト紹介）
+    - `BlogCard` コンポーネントを修正し、新しいURL構造をサポート
+      - propsインターフェースにhrefプロパティを追加
+      - デフォルト値を `/articles/${slug}` に変更
+      - ShareButtonコンポーネントのURLも更新
+    - コンパイルエラーを修正
+      - `page.tsx`でのBlogCardコンポーネント使用方法を修正（`post={post}`形式から展開したプロパティ形式へ）
+      - TagCloudコンポーネントのプロパティ名を`maxItems`から`limit`に修正
+    - URL参照を含むテストも適切に更新
+  - 効果:
+    - 直感的で意味のあるURL構造の実現
+    - トップページのユーザー体験と情報アーキテクチャの改善
+    - URL構造の一貫性向上によるSEOへの好影響
+- Memory Bankも最新化し、変更を記録
+- プロジェクト全体で47テストが正常に通過
+
 ## 2025-05-15 OG画像生成のフォント読み込み改善とクリーンアップ
 
 - OG画像生成の環境依存問題を解決
