+++ matome-blog/src/app/page.tsx
@@ -1,9 +1,13 @@
-import { BlogList } from "@/components/blog/BlogList";
+import { BlogCard } from "@/components/blog/BlogCard";
 import { prisma } from "@/lib/db/prisma";
 import { PostWithReadingTime } from "@/lib/db/types";
-import { notFound } from "next/navigation";
 import { Metadata } from "next";
 import { JsonLd } from "@/components/blog/JsonLd";
+import Link from "next/link";
+import { Button } from "@/components/ui/button";
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
+import { TagCloud } from "@/components/blog/TagComponents";
+import { getAllTags } from "@/lib/tags";
 
 // Generate metadata for home page
 export const metadata: Metadata = {
@@ -23,27 +27,21 @@ export const metadata: Metadata = {
   },
 };
 
-interface HomeProps {
-  searchParams?: { [key: string]: string | string[] | undefined };
-}
+export default async function Home() {
+  // 新着記事（最新5件）を取得
+  const latestPosts = await prisma.post.findMany({
+    orderBy: { publishedAt: "desc" },
+    include: { tags: true },
+    take: 5,
+  });
 
-export default async function Home({ searchParams }: HomeProps) {
-  const params = searchParams ? await searchParams : {};
-  const page = typeof params.page === "string" ? parseInt(params.page, 10) : 1;
-  const limit = 9;
-  const skip = (page - 1) * limit;
-  const [posts, total] = await Promise.all([
-    prisma.post.findMany({
-      orderBy: { publishedAt: "desc" },
-      include: { tags: true },
-      skip,
-      take: limit,
-    }),
-    prisma.post.count(),
-  ]);
+  // すべての記事を取得してタグクラウド用のデータ準備
+  const allPosts = await prisma.post.findMany({
+    include: { tags: true },
+  });
+  const tags = getAllTags(allPosts);
 
-  const totalPages = Math.max(1, Math.ceil(total / limit));
-  if (page < 1 || (totalPages > 0 && page > totalPages)) notFound();  return (
+  return (
     <>
       <JsonLd
         data={{
@@ -62,15 +60,84 @@ export default async function Home({ searchParams }: HomeProps) {
         }}
       />
       <main className="container mx-auto px-4 py-8">
-        <h1 className="text-4xl font-bold mb-8">YouTube Summary Blog</h1>
-        <BlogList
-          posts={posts}
-          currentPage={page}
-          totalPages={totalPages}
-          baseUrl="/"
-          showPagerTop={true}
-          showPagerBottom={true}
-        />
+        {/* ヒーローセクション */}
+        <section className="mb-12 text-center">
+          <h1 className="text-4xl md:text-5xl font-bold mb-4">YouTube Summary Blog</h1>
+          <p className="text-xl text-muted-foreground max-w-3xl mx-auto mb-6">
+            人気のYouTube動画をテキストで効率的に理解できるサマリーブログです。
+            気になる動画の内容をすばやく把握したい方におすすめです。
+          </p>
+          <div className="flex flex-wrap justify-center gap-4">
+            <Button asChild size="lg">
+              <Link href="/blog">すべての記事を見る</Link>
+            </Button>
+            <Button asChild variant="outline" size="lg">
+              <Link href="/tags">タグから探す</Link>
+            </Button>
+          </div>
+        </section>
+
+        {/* 新着記事 */}
+        <section className="mb-12">
+          <div className="flex justify-between items-center mb-6">
+            <h2 className="text-3xl font-bold">新着記事</h2>
+            <Button asChild variant="ghost">
+              <Link href="/blog">もっと見る →</Link>
+            </Button>
+          </div>
+          <div className="grid gap-6 grid-cols-1 md:grid-cols-2 lg:grid-cols-3">            {latestPosts.map((post) => (
+              <BlogCard
+                key={post.id}
+                title={post.title}
+                excerpt={post.excerpt}
+                slug={post.slug}
+                publishedAt={post.publishedAt}
+                videoId={post.videoId}
+                readingTimeMinutes={(post as PostWithReadingTime).readingTimeMinutes}
+                tags={post.tags}
+                href={`/articles/${post.slug}`}
+              />
+            ))}
+          </div>
+        </section>
+
+        {/* タグクラウド */}
+        <section className="mb-12">
+          <Card>
+            <CardHeader>
+              <CardTitle>人気のタグ</CardTitle>
+              <CardDescription>興味のあるトピックから記事を探す</CardDescription>
+            </CardHeader>
+            <CardContent>
+              <TagCloud tags={tags} limit={20} className="justify-center" />
+              <div className="text-center mt-4">
+                <Button asChild variant="outline">
+                  <Link href="/tags">すべてのタグを見る</Link>
+                </Button>
+              </div>
+            </CardContent>
+          </Card>
+        </section>
+
+        {/* サイト紹介 */}
+        <section className="mb-12">
+          <Card>
+            <CardHeader>
+              <CardTitle>このブログについて</CardTitle>
+              <CardDescription>YouTube動画の要約コンテンツプラットフォーム</CardDescription>
+            </CardHeader>
+            <CardContent>
+              <p className="mb-4">
+                このブログは、YouTube動画の要約コンテンツを効果的に配信するためのプラットフォームです。
+                検索エンジン最適化（SEO）とAIによる情報解析のしやすさを重視し、MDX形式で管理された記事を
+                高速かつ快適に閲覧できます。
+              </p>
+              <Button asChild>
+                <Link href="/about">詳しく見る</Link>
+              </Button>
+            </CardContent>
+          </Card>
+        </section>
       </main>
     </>
   );
